+++ cmd/uhppote-cli/main.go
@@ -55,6 +55,7 @@ var cli = []commands.Command{
 	&commands.SetPCControlCmd,
 	&commands.SetInterlockCmd,
 	&commands.ActivateKeypadsCmd,
+	&commands.SetSuperPasswordsCmd,
 	&commands.ListenCmd,
 }